Fix KMC tools max thread count.
SBWT construction was sometimes hanging on Slurm clusters. We think that this could be due to how
KMC sort uses by default a number of threads equal to thread::hardware_concurrency().
This might cause problems if the number of cores Slurm allocates to the process is much smaller
than the number of cores on the node. This patch sets the number of threads equal to the
user-defined number.